fix: return value should be int

+++ app/Http/Livewire/Dealer/Audit/BodyShop/IndexItem.php
@@ -77,7 +77,7 @@ class IndexItem extends Component
 
     public function remediationProgress(): int
     {
-        return $this->bodyShopAudit->violation_count === 0 ? 0 : round($this->bodyShopAudit->remediation_count / $this->bodyShopAudit->violation_count * 100);
+        return $this->bodyShopAudit->violation_count === 0 ? 0 : (int) round($this->bodyShopAudit->remediation_count / $this->bodyShopAudit->violation_count * 100);
     }
 
     public function delete(): void
